Add new error type for serde_dhall
